+++ tests/regression/test_graph_scope_resolve.py
@@ -13,8 +13,8 @@ def test_scope_resolve_fallback():
     with graph.scope() as scope:
         u2 = scope.resolve(UserService)
 
-    assert UserService in graph._resolved_instances
-    assert UserService not in scope._resolved_instances
+    assert UserService in graph._resolved_singletons
+    assert UserService not in scope._resolved_singletons
 
     assert u is u2
 
@@ -32,8 +32,8 @@ def test_scope_resouce_fallback():
     with graph.scope() as scope:
         u2 = scope.resolve(UserService)
 
-    assert UserService not in graph._resolved_instances
-    assert UserService in scope._resolved_instances
+    assert UserService not in graph._resolved_singletons
+    assert UserService in scope._resolved_singletons
 
     assert u is not u2