+++ tests/regression/test_slots.py
@@ -1,45 +1,26 @@
 import pytest
 
 from ididi._node import Dependencies, Dependency, DependentNode
-from ididi.graph import AsyncScope, Graph, ScopeManager, SyncScope
+from ididi.graph import Graph
 
 
-def test_graph_ds_slots():
+async def test_graph_ds_slots():
     dg = Graph()
-    sm = ScopeManager(1, 2, 3, 4)
-    asc = AsyncScope(
-        graph_resolutions=1,
-        graph_singletons=2,
-        name=2,
-        pre=3,
-        nodes=4,
-        resolved_nodes=5,
-        type_registry=6,
-        ignore=7,
-    )
-
-    sc = SyncScope(
-        graph_resolutions=1,
-        graph_singletons=9,
-        name=2,
-        pre=3,
-        nodes=4,
-        resolved_nodes=5,
-        type_registry=6,
-        ignore=7,
-    )
-
     with pytest.raises(AttributeError):
         dg.__dict__
+
+    sm = dg.scope()
 
     with pytest.raises(AttributeError):
         sm.__dict__
 
-    with pytest.raises(AttributeError):
-        asc.__dict__
+    async with sm as asc:
+        with pytest.raises(AttributeError):
+            asc.__dict__
 
-    with pytest.raises(AttributeError):
-        sc.__dict__
+    with sm as sc:
+        with pytest.raises(AttributeError):
+            sc.__dict__
 
 
 def test_node_ds_slots():