+++ CHANGELOG.md
@@ -949,8 +949,16 @@ and such override won't affect others.
 
 ## version 1.3.5
 
-a quick bug fix, where in 1.3.4, registered_singleton is not shared between graph and scope.
+- a quick bug fix, where in 1.3.4, registered_singleton is not shared between graph and scope.
 
 The general rule is that scope can access registered singletons and resolved instances but not vice versa.
+
+
+
+## version 1.3.6
+
+
+- Fix: `Graph.entry` no longer uses existing scope, instead, always create a new scope
+